+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+* Changed `Repo.fetch/2` and `Repo.fetch/3` to work with primary key with name other than `id`.
+
 ## [0.13.2] - 2022-12-19
 
 * Add `error_tag` option to `BitcrowdEcto.Repo.fetch_by/3` and friends to allow overriding the queryable in the error result.